# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Municipalities reports API",
        "description": "API for municipalities reports",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://backend.munireports.test/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/v1/reports": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Get list of reports",
                "description": "Returns list of reports",
                "operationId": "getReportsList",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Status values that needed to be considered for filter",
                        "required": false,
                        "explode": true,
                        "schema": {
                            "type": "string",
                            "default": "open",
                            "enum": [
                                "open",
                                "in_progress",
                                "resolved",
                                "closed"
                            ]
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Limit number of results",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFound"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Reports"
                ],
                "summary": "Create a new report",
                "operationId": "createReport",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateReportRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Report created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CreateReportRequest"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BadRequest"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Unauthenticated"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Unauthorized"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/v1/reports/{id}": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Get report information",
                "description": "Returns report data",
                "operationId": "getReportById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Report id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Unauthenticated"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Unauthorized"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Reports"
                ],
                "summary": "Update existing report",
                "description": "Returns updated report data",
                "operationId": "updateReport",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Report id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateReportRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Report updated successfully!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UpdateReportRequest"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BadRequest"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Unauthenticated"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Unauthorized"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Reports"
                ],
                "summary": "Delete existing report",
                "description": "Deletes a record and returns no content",
                "operationId": "deleteReport",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Report id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Unauthenticated"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Unauthorized"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/v1/reports/{id}/votes": {
            "post": {
                "tags": [
                    "Reports"
                ],
                "summary": "Vote on a report",
                "description": "Vote on a report",
                "operationId": "vote",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of report to vote on",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Vote object that needs to be added to the store",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "vote": {
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Vote response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {}
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "unexpected error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "An error occurred"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "BadRequest": {
                "description": "Bad request",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Bad request"
                    }
                },
                "type": "object"
            },
...
            "NotFound": {
                "description": "La ressource n'existe pas",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "La ressource demandée n'a pas pu être trouvée"
                    }
                },
                "type": "object"
            },
            "Unauthenticated": {
                "description": "Unauthenticated",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Unauthenticated"
                    }
                },
                "type": "object"
            },
            "Unauthorized": {
                "description": "Unauthorized",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Unauthorized"
                    }
                },
                "type": "object"
            },
...
            "CreateReportRequest": {
                "required": [
                    "title",
                    "description",
                    "location",
                    "category"
                ],
                "properties": {
                    "title": {
                        "description": "Get the validation rules that apply to the request.",
                        "type": "string",
                        "example": "Report title"
                    },
                    "description": {
                        "type": "string",
                        "example": "Report description"
                    },
                    "location": {
                        "type": "array",
                        "items": {
                            "type": "number",
                            "format": "float"
                        },
                        "example": [
                            -6.175392,
                            106.827153
                        ]
                    },
                    "category": {
                        "type": "string",
                        "example": "Report category"
                    },
                    "medias": {
                        "description": "Report medias",
                        "type": "array",
                        "items": {
                            "type": "file",
                            "format": "binary"
                        }
                    }
                },
                "type": "object"
            },
            "UpdateReportRequest": {
                "properties": {
                    "title": {
                        "description": "Get the validation rules that apply to the request.",
                        "type": "string",
                        "example": "Report title"
                    },
                    "description": {
                        "type": "string",
                        "example": "Report description"
                    },
                    "location": {
                        "type": "string",
                        "example": "Report location"
                    },
                    "category": {
                        "type": "string",
                        "example": "Report category"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}